backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.9
pdfplumber==0.11.0
PyMuPDF==1.24.3
openai==1.30.1
anthropic==0.28.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
aiofiles==23.2.1
httpx==0.27.0